Set permissions for deleting comments

+++ comments/views.py
@@ -1,6 +1,7 @@
 from django.contrib import messages
 from django.contrib.contenttypes.models import ContentType
-from django.http import HttpResponseRedirect
+from django.http import HttpResponse
+from django.http import HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404
 
 from .models import Comment
@@ -8,7 +9,16 @@ from .forms import CommentForm
 
 
 def comment_delete(request, pk):
-    comment = get_object_or_404(Comment, pk=pk)
+    try:
+        comment = Comment.objects.get(pk=pk)
+    except:
+        raise Http404
+
+    if comment.user != request.user:
+        response = HttpResponse("You don't have permission to do this")
+        response.status_code = 403
+        return response
+
     if request.method == 'POST':
         messages.success(request, "Comment successfully deleted")
         comment.delete()
@@ -21,11 +31,18 @@ def comment_delete(request, pk):
 
 
 def comment_thread(request, pk):
-    comment = get_object_or_404(Comment, pk=pk)
+    try:
+        comment = Comment.objects.get(pk=pk)
+    except:
+        raise Http404
+
     initial_data = {
         "content_type": comment.content_type,
         "object_id": comment.object_id,
     }
+
+    if not comment.is_parent:
+        comment = comment.parent
 
     form = CommentForm(request.POST or None, initial=initial_data)
